Fix react-native/metro bundler warning
When using Zod in a react-native app, we see the following warning when running the bundler:
```
warn Package zod has been ignored because it contains invalid configuration. Reason: Package subpath './package.json' is not defined by "exports" in <path>/node_modules/zod/package.json
```

The problem is described in https://github.com/react-native-community/cli/issues/1168 and that is where it should be properly fixed. In the meantime (that issue has been open for some time), the warning can be fixed in affected packages by adding an export for `package.json`.